Домашка по JavaCore.Урок №3.
